hope it works maybe

+++ project/camera.h
@@ -12,7 +12,7 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-uint8_t processCameraData(uint16_t *adcData, size_t length);
+uint8_t processCameraData(uint16_t *adcData, uint8_t length);
 void getSmoothedData(uint16_t *dest, size_t length);
 void getDerivData(uint16_t *dest, size_t length);
